Add authentication and authorization to Like controller with tests

+++ app/controllers/likes_controller.rb
@@ -1,4 +1,6 @@
 class LikesController < ApplicationController
+  before_action :require_login, only:[:create, :destroy]
+
   def create
     secret = Secret.find(params[:secret_id])
     Like.create(user: current_user, secret: secret)
@@ -6,8 +8,12 @@ class LikesController < ApplicationController
   end
 
   def destroy
-    like = Like.find_by(user: current_user, secret: params[:id])
-    like.destroy
-    redirect_to "/secrets"
+    like = Like.find_by(user: current_user, secret: params[:secret_id])
+    if like != nil && like.user == current_user
+      like.destroy
+      redirect_to "/secrets"
+    else
+      redirect_to "/users/#{current_user.id}"
+    end
   end
 end
